# packages/frontend/src/redux/schema/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts": {
      "post": {
        "operationId": "Accounts_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAccountDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "properties": { "payload": { "type": "string" } } }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "get": {
        "operationId": "Accounts_findAllByUser",
        "parameters": [
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/AccountType" }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AccountDto" }
                }
              }
            }
          }
        },
        "tags": ["Accounts"]
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "operationId": "Accounts_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return account by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountDto" }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "patch": {
        "operationId": "Accounts_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAccountDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountDto" }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "delete": {
        "operationId": "Accounts_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Accounts"]
      }
    },
    "/api/accounts/{id}/balance-history": {
      "get": {
        "operationId": "Accounts_getAccountBalanceHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return account balance history by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountBalanceHistoryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Accounts"]
      }
    },
...
    "/api/incomes": {
      "get": {
        "operationId": "Incomes_findAllByUser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/IncomeListItemDto" }
                }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "post": {
        "operationId": "Incomes_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateIncomeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDetailsDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      }
    },
    "/api/incomes/{id}": {
      "get": {
        "operationId": "Incomes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDetailsDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "patch": {
        "operationId": "Incomes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateIncomeDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IncomeDetailsDto" }
              }
            }
          }
        },
        "tags": ["Incomes"]
      },
      "delete": {
        "operationId": "Incomes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Incomes"]
      }
    },
...
  },
  "info": {
    "title": "Financer",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AccountType": {
        "type": "string",
        "enum": [
          "CASH",
          "SAVINGS",
          "INVESTMENT",
          "CREDIT",
          "LOAN",
          "LONG_TERM_SAVINGS",
          "PRE_ASSIGNED_CASH"
        ]
      },
      "AccountDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "name": { "type": "string" },
          "type": { "allOf": [{ "$ref": "#/components/schemas/AccountType" }] },
          "balance": { "type": "number" },
          "currentDateBalance": { "type": "number", "nullable": true },
          "userId": { "type": "string" },
          "isDeleted": { "type": "boolean" }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "name",
          "type",
          "balance",
          "currentDateBalance",
          "userId",
          "isDeleted"
        ]
      },
...
      "TransactionType": {
        "type": "string",
        "enum": ["INCOME", "EXPENSE", "TRANSFER"]
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "allOf": [{ "$ref": "#/components/schemas/AccountType" }] },
          "balance": { "type": "number" }
        },
        "required": ["name", "type", "balance"]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "allOf": [{ "$ref": "#/components/schemas/AccountType" }] },
          "balance": { "type": "number" }
        }
      },
      "AccountBalanceHistoryDto": {
        "type": "object",
        "properties": {
          "date": { "format": "date-time", "type": "string" },
          "amount": { "type": "number" },
          "balance": { "type": "number" }
        },
        "required": ["date", "amount", "balance"]
      },
      "TransactionListItemCategoryDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      },
...
      "TransactionDetailCategoryDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "amount": { "type": "number" },
          "name": { "type": "string" },
          "path": { "type": "string" }
        },
        "required": ["id", "description", "amount", "name", "path"]
      },
...
      "CreateTransactionCategoryMappingWithoutTransactionDto": {
        "type": "object",
        "properties": {
          "description": { "type": "string", "nullable": true },
          "categoryId": { "type": "string" },
          "amount": { "type": "number" }
        },
        "required": ["description", "categoryId", "amount"]
      },
...
      "IncomeDetailsDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "toAccount": { "type": "string", "nullable": true },
          "toAccountName": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" },
          "type": {
            "description": "Type of the transaction",
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDetailCategoryDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "toAccount",
          "toAccountName",
          "isRecurring",
          "type",
          "categories"
        ]
      },
      "IncomeListItemDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "toAccount": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" },
          "type": {
            "description": "Type of the transaction",
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionListItemCategoryDto"
            }
          }
        },
        "required": [
          "id",
          "amount",
          "description",
          "date",
          "toAccount",
          "isRecurring",
          "type",
          "categories"
        ]
      },
      "CreateIncomeDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "toAccount": { "type": "string", "nullable": true },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": ["amount", "description", "date"]
      },
      "UpdateIncomeDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "toAccount": { "type": "string", "nullable": true },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        }
      },
...
    }
  }
}